backend/pyproject.toml
[project]
name = "travel-assistant-backend"
version = "0.1.0"
description = "Travel Assistant Agent Backend with MCP Client"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "httpx[socks]>=0.28.1",
    "mcp>=1.6.0",
    "openai>=1.71.0",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "googlemaps>=4.10.0",
    "aiofiles>=23.2.1",
    "pandas>=2.1.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "ipython>=9.1.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 